Renamed MockFileSystem into ComposableFileSystem, simplified.
